gateway/pyproject.toml
[project]
name = "presenton-gateway"
version = "0.1.0"
description = "API Gateway + MCP Gateway"
requires-python = ">=3.11,<3.13"
dependencies = [
    "presenton-shared",
    "fastapi[standard]>=0.116.0",
    "PyJWT[crypto]>=2.9.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api_gateway", "mcp_gateway"]

[tool.uv.sources]
presenton-shared = { workspace = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "cryptography>=43.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "../.."]
